+++ src/App.jsx
@@ -1,7 +1,7 @@
 import './App.css';
 import React from 'react';
 import { ThemeProvider } from 'styled-components';
-import Login from './pages/login/Login';
+import Login from './pages/login/Login.page';
 import theme from './theme';
 
 function App() {
