+++ src/theme.js
@@ -3,6 +3,7 @@ const theme = {
     background1: '#ced4da',
     colorPrimary: '#55c57a',
     colorPrimaryDark: '#40916c',
+    colorError: '#d00000',
   },
 };
